Revert "fluent viewport pos (float world pos)"
This reverts commit debda4f2dfaa434f4be5e33a9c0065cb0da36181.

Don't like it. Too much stuff expects the LX single res,
and is drawn like that, like projectiles, etc.
Also, worm is warping up/down because of simulation and
because worm pos is float, which was not visible before
and should not be visible.

+++ include/CViewport.h
@@ -85,7 +85,7 @@ private:
 
 	int		VirtWidth, VirtHeight;
 
-	float  	WorldX, WorldY;
+	int  	WorldX, WorldY;
 
     CVec    curPos, tgtPos;
